// src/types/nusmods.ts
export type ModuleCode = string; // E.g. "CS3216"
export type ModuleTitle = string;
// ...
export type ModuleCondensed = Readonly<{
  moduleCode: ModuleCode;
  title: ModuleTitle;
  semesters: number[];
}>;